Fix typo
Originally, the function for getting the hash of a node tuple was
calling some crazy dictionary, but I removed that usage because it can
be calculated quickly and cheaply enough on the fly (this make
programming  a lot more simple)

I removed the cache that you put in and changed the function to
calculate it directly. Obviously its dumb to wrap a static method with
an instance method, so I put a warning so we can remove all references
to it in the future.

@ddomingof